Fix chapter 9 pentagon images
The polygons and curved-polygons image files for SVG format were
correct, but the PDF format showed a rhombus instead of a pentagon.
I've used the online converter at fileformat.info to generate PDFs
from the SVGs. The PDFs have the same dimensions as the old ones.
You can verify by building the PDF format of the book and checking
figures 9.2 and 9.3.